Refactor enrollment types: move EnrollmentStatus and EnrollmentDetails to a new file
