--- ee/agent/types/status.go
@@ -1,10 +1,0 @@
-package types
-
-type EnrollmentStatus string
-
-const (
-	NoEnrollmentKey EnrollmentStatus = "no_enrollment_key"
-	Unenrolled      EnrollmentStatus = "unenrolled"
-	Enrolled        EnrollmentStatus = "enrolled"
-	Unknown         EnrollmentStatus = "unknown"
-)
